Brings the fuel consumption, tire wear, and pit-stops release (plus 500-lap
support and related changes) onto the custom-vehicles-mp branch.

Three conflicts, all in the multiplayer drive-session construction seam where
both features extended the same constructor/factory/call site:
- Session/Core/Core.cs, Game/Drive/SessionFactory.cs, Game/Multiplayer/Race/Runtime.cs
Resolved by combining both parameter sets: main's RacePhysicsToggles alongside
this branch's resolveRemoteVehicleFile/Name, ordered physicsToggles then
resolvePlayerName then the two remote-vehicle resolvers (optional args last).

Both projects build clean; full suite 529/529 pass.

+++ top_speed_net/TopSpeed.Shared/Protocol/TrackPackages/Manifest.cs
@@ -9,6 +9,6 @@ namespace TopSpeed.Protocol
         public string Hash = string.Empty;
         public string DefaultWeatherProfileId = TrackWeatherProfile.DefaultProfileId;
         public TrackAmbience Ambience;
-        public byte Laps;
+        public int Laps;
     }
 }
